Modulariza módulo de calidad a app dedicada
