+++ backend/calidad/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class CalidadConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'backend.calidad'
+    verbose_name = 'Calidad'
